backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.3.0
Pillow==12.0.0
stripe==7.4.0
requests==2.31.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
drf-yasg==1.21.7
pymongo==4.6.0